Refactor move nodes to own folder
