+++ src/compiler/nodes/interfaces.ts
@@ -1,0 +1,9 @@
+import { TreeNode } from "./";
+
+export interface IVisitable {
+  Accept: (visitor: IVisitor) => void;
+}
+
+export interface IVisitor {
+  VisitNode: (node: TreeNode) => void;
+}
